+++ src/solutions/task_1.js
