+++ src/solutions/task_2.js
